feat(Dashboard): migrate to declarative manifest widgets (Option A) — closes #536
Replace the custom DashboardView component (which wrapped CnDashboardPage
inside a CnWidgetWrapper inside the manifest dashboard route — triple
nesting) with a fully-declarative manifest dashboard entry:

- src/manifest.json Dashboard page is now type:"dashboard" with
  config.widgets[] + config.layout[] + slots[] + actionsComponent
- 8 self-contained widget components under src/views/dashboard/widgets/,
  each fetching its own data via the new dashboardData.js service that
  memoises one in-flight promise per dataset (5-min TTL) so the 4 KPI
  widgets don't trigger 4 duplicate /leads + /pipelines requests
- 4 KPIs (OpenLeads, OpenRequests, PipelineValue, Overdue) ship as
  custom widgets wrapping <CnStatsBlock> rather than declarative
  type:"stats-block" entries — the values require cross-schema
  joins (e.g. overdue = leads filtered against pipeline-stage
  isClosed Set) that the stats-block dataSource aggregate shorthand
  cannot express
- 4 lower panels (RequestsByStatus, Complaints, MyWork, ClientOverview)
  are extracted into their own components with no shared parent state
- DashboardHeaderActions hosts the 3 New X buttons + Refresh, wired via
  the page's actionsComponent
- src/views/Dashboard.vue is deleted
- src/registry.js + src/customComponents.js drop the DashboardView entry
  and register the 9 new widget components

Nesting reduction: 3 wrappers (manifest dashboard → CnWidgetWrapper →
DashboardView → inner CnDashboardPage) → 1 wrapper (CnDashboardPage,
mounted directly by CnPageRenderer from the manifest).
