+++ src/customComponents.js
@@ -3,14 +3,14 @@
 //
 // Custom-component registry for pipelinq's manifest-driven app shell.
 //
-// Every entry here is the "escape hatch" — pages that don't fit one
-// of the manifest's built-in types/widgets. Keep this file focused.
+// Every entry here is the "escape hatch" — pages and widgets that don't
+// fit one of the manifest's built-in types. Keep this file focused.
 // Adding entries requires explicit justification in the design doc;
 // removing them (by migrating to a built-in type) is the right
 // direction.
 //
 // Resolution order at runtime:
-//   1. Built-in page types          (CnIndexPage, CnDetailPage, …)
+//   1. Built-in page types          (CnIndexPage, CnDetailPage, CnDashboardPage, …)
 //   2. Built-in widget types        (version-info, register-mapping, …)
 //   3. customComponents (this file) ← consumer-injected components
 //
@@ -18,10 +18,24 @@
 //   - openspec/changes/pipelinq-manifest-v1/design.md
 //   - hydra/openspec/architecture/adr-024-app-manifest.md
 
-// --- Genuine exceptions (no abstract analogue). ---
-import DashboardView from './views/Dashboard.vue'
+// --- Pipeline board — bespoke kanban (lib gap: no kanban page type). ---
 import PipelineBoardView from './views/pipeline/PipelineBoard.vue'
 import MyWorkView from './views/MyWork.vue'
+
+// --- Dashboard (manifest type:"dashboard") — header actions and
+//     per-widget slot components. The Dashboard route itself is rendered
+//     by CnDashboardPage; this file only exposes the 8 widget components
+//     and the header actions component that the manifest's `slots` and
+//     `actionsComponent` fields reference by name. ---
+import DashboardHeaderActions from './views/dashboard/DashboardHeaderActions.vue'
+import OpenLeadsKpiWidget from './views/dashboard/widgets/OpenLeadsKpiWidget.vue'
+import OpenRequestsKpiWidget from './views/dashboard/widgets/OpenRequestsKpiWidget.vue'
+import PipelineValueKpiWidget from './views/dashboard/widgets/PipelineValueKpiWidget.vue'
+import OverdueKpiWidget from './views/dashboard/widgets/OverdueKpiWidget.vue'
+import RequestsByStatusWidget from './views/dashboard/widgets/RequestsByStatusWidget.vue'
+import ComplaintsWidget from './views/dashboard/widgets/ComplaintsWidget.vue'
+import MyWorkWidget from './views/dashboard/widgets/MyWorkWidget.vue'
+import ClientOverviewWidget from './views/dashboard/widgets/ClientOverviewWidget.vue'
 
 // --- Queues — bespoke routing-rule editor (lib gap: routing-rules widget). ---
 import QueueListView from './views/queues/QueueList.vue'
@@ -59,10 +73,20 @@ import SyncSettingsView from './views/sync/SyncSettings.vue'
 //     OpenRegister's github-issue-proxy). See ConductionNL/hydra#251. ---
 
 export default {
-	// Genuine exceptions
-	DashboardView,
+	// Pipeline + MyWork
 	PipelineBoardView,
 	MyWorkView,
+
+	// Dashboard widget components (resolved via Dashboard page's `slots` map)
+	DashboardHeaderActions,
+	OpenLeadsKpiWidget,
+	OpenRequestsKpiWidget,
+	PipelineValueKpiWidget,
+	OverdueKpiWidget,
+	RequestsByStatusWidget,
+	ComplaintsWidget,
+	MyWorkWidget,
+	ClientOverviewWidget,
 
 	// Queues
 	QueueListView,
