+++ src/registry.js
@@ -3,14 +3,14 @@
 //
 // V2 component registry for pipelinq.
 //
-// Every entry here corresponds to a manifest `type: "custom"` page. The
-// registry maps the string key used in the manifest to a `{ kind, component }`
-// entry so CnAppRoot can resolve the component at render time.
+// Every entry here corresponds to a manifest `type: "custom"` page, a
+// dashboard widget rendered via a `slots` mapping, or a
+// `headerComponent` / `actionsComponent` override on a typed page.
 //
 // Recognised kinds: page, modal, widget, form-field, cell-renderer
 //
 // Resolution order (v2 renderer):
-//   1. Built-in page types   (CnIndexPage, CnDetailPage, …)
+//   1. Built-in page types   (CnIndexPage, CnDetailPage, CnDashboardPage, …)
 //   2. Built-in widget types (version-info, register-mapping, …)
 //   3. registry (this file)  ← consumer-injected components
 //   4. customComponents      ← v1 fallback, kept during transition
@@ -19,9 +19,21 @@
 //   - openspec/changes/pipelinq-manifest-v1/design.md
 //   - hydra/openspec/architecture/adr-036-manifest-v2.md
 
-// --- Genuine exceptions: no abstract manifest analogue. ---
-import DashboardView from './views/Dashboard.vue'
+// --- MyWork — bespoke per-user surface mixing tasks + leads + requests. ---
 import MyWorkView from './views/MyWork.vue'
+
+// --- Dashboard (manifest-driven type:"dashboard") — header actions and
+//     per-widget slot components. The page itself is rendered by
+//     CnDashboardPage from `config.widgets[]` + `config.layout[]`. ---
+import DashboardHeaderActions from './views/dashboard/DashboardHeaderActions.vue'
+import OpenLeadsKpiWidget from './views/dashboard/widgets/OpenLeadsKpiWidget.vue'
+import OpenRequestsKpiWidget from './views/dashboard/widgets/OpenRequestsKpiWidget.vue'
+import PipelineValueKpiWidget from './views/dashboard/widgets/PipelineValueKpiWidget.vue'
+import OverdueKpiWidget from './views/dashboard/widgets/OverdueKpiWidget.vue'
+import RequestsByStatusWidget from './views/dashboard/widgets/RequestsByStatusWidget.vue'
+import ComplaintsWidget from './views/dashboard/widgets/ComplaintsWidget.vue'
+import MyWorkWidget from './views/dashboard/widgets/MyWorkWidget.vue'
+import ClientOverviewWidget from './views/dashboard/widgets/ClientOverviewWidget.vue'
 
 // --- Kanban board (lib gap: no kanban/board page type). ---
 import PipelineBoardView from './views/pipeline/PipelineBoard.vue'
@@ -65,16 +77,64 @@ import SyncSettingsView from './views/sync/SyncSettings.vue'
  * @type {Record<string, { kind: string, component: object, _note?: string }>}
  */
 const registry = {
-	// --- Genuine exceptions: no abstract manifest analogue. ---
-	DashboardView: {
-		kind: 'page',
-		component: DashboardView,
-		_note: 'Bespoke multi-widget dashboard with gridstack layout; lib gap: no declarative dashboard page type with grid-aware widget placement.',
-	},
+	// --- MyWork — multi-entity user dashboard. ---
 	MyWorkView: {
 		kind: 'page',
 		component: MyWorkView,
 		_note: 'Personalised work surface mixing tasks + leads + requests for the current user; no single-entity typed page captures multi-entity user dashboard.',
+	},
+
+	// --- Dashboard widgets (rendered as #widget-{id} slots inside
+	//     CnDashboardPage via the manifest Dashboard page's `slots` map).
+	//     KPI widgets compute their counts from multiple cross-schema
+	//     fetches (e.g. overdue = leads filtered against pipeline
+	//     isClosed stages) that the declarative `stats-block` +
+	//     `dataSource` shorthand can't express, so they ship as small
+	//     custom widget components instead. ---
+	DashboardHeaderActions: {
+		kind: 'widget',
+		component: DashboardHeaderActions,
+		_note: 'Dashboard header buttons (New Lead / Request / Client + Refresh) wired as the Dashboard page actionsComponent.',
+	},
+	OpenLeadsKpiWidget: {
+		kind: 'widget',
+		component: OpenLeadsKpiWidget,
+		_note: 'KPI card for open leads (leads minus those in pipeline stages flagged isClosed). Renders <CnStatsBlock>.',
+	},
+	OpenRequestsKpiWidget: {
+		kind: 'widget',
+		component: OpenRequestsKpiWidget,
+		_note: 'KPI card for open requests (status new or in_progress). Renders <CnStatsBlock>.',
+	},
+	PipelineValueKpiWidget: {
+		kind: 'widget',
+		component: PipelineValueKpiWidget,
+		_note: 'KPI card for total open-lead value in EUR. Renders <CnStatsBlock>.',
+	},
+	OverdueKpiWidget: {
+		kind: 'widget',
+		component: OverdueKpiWidget,
+		_note: 'KPI card for overdue leads + stale requests. Renders <CnStatsBlock>.',
+	},
+	RequestsByStatusWidget: {
+		kind: 'widget',
+		component: RequestsByStatusWidget,
+		_note: 'Horizontal bar chart of requests grouped by status. Standalone widget — fetches its own data.',
+	},
+	ComplaintsWidget: {
+		kind: 'widget',
+		component: ComplaintsWidget,
+		_note: 'Open / overdue / status breakdown of complaints. Wraps the existing ComplaintsOverviewWidget with a self-contained fetch.',
+	},
+	MyWorkWidget: {
+		kind: 'widget',
+		component: MyWorkWidget,
+		_note: 'Top-5 list of leads + requests assigned to the current user, sorted by overdue → priority → due date.',
+	},
+	ClientOverviewWidget: {
+		kind: 'widget',
+		component: ClientOverviewWidget,
+		_note: 'Top-5 recent clients with a view-all link to ClientList.',
 	},
 
 	// --- Kanban board. ---
