+++ src/views/dashboard/widgets/OpenLeadsKpiWidget.vue
@@ -1,0 +1,38 @@
+<template>
+	<CnStatsBlock
+		:title="t('pipelinq', 'Open Leads')"
+		:count="count"
+		:count-label="t('pipelinq', 'leads')"
+		:icon="TrendingUp"
+		variant="primary"
+		horizontal
+		:route="{ name: 'Leads', query: { status: 'open' } }" />
+</template>
+
+<script>
+import { CnStatsBlock } from '@conduction/nextcloud-vue'
+import TrendingUp from 'vue-material-design-icons/TrendingUp.vue'
+import { getLeads, getPipelines, getClosedStageNames } from '../../../services/dashboardData.js'
+
+export default {
+	name: 'OpenLeadsKpiWidget',
+	components: {
+		CnStatsBlock,
+	},
+	data() {
+		return {
+			TrendingUp,
+			count: 0,
+		}
+	},
+	async mounted() {
+		try {
+			const [leads, pipelines] = await Promise.all([getLeads(), getPipelines()])
+			const closed = getClosedStageNames(pipelines)
+			this.count = leads.filter(l => !closed.has(l.stage)).length
+		} catch (err) {
+			console.error('OpenLeadsKpiWidget fetch error:', err)
+		}
+	},
+}
+</script>
